Modify index.html.erb

+++ db/migrate/20181204120458_add_star_to_posts.rb
@@ -1,8 +1,8 @@
 class AddStarToPosts < ActiveRecord::Migration[5.2]
   def change
-    add_column :posts, :star_full, :integer
-    add_column :posts, :star_seat, :integer
-    add_column :posts, :star_ca, :integer
-    add_column :posts, :star_before, :integer
+    add_column :posts, :star_full, :float
+    add_column :posts, :star_seat, :float
+    add_column :posts, :star_ca, :float
+    add_column :posts, :star_before, :float
   end
 end
